requirements.txt
fastapi==0.99.1
fastapi-cli==0.0.3
langchain==0.1.19
langchain-chroma==0.1.0
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-vertexai==1.0.1
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langchainhub==0.1.15
pandas==2.2.1
Pillow==10.0.1
pillow_heif==0.16.0
openai==1.27.0
SQLAlchemy==2.0.23
sqlitedict==2.1.0
streamlit==1.34.0
streamlit-chat==0.1.1